React overhaul base
