+++ src/App.css
@@ -1,0 +1,238 @@
+@import url("https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&display=swap");
+
+:root {
+  /* Main site colors */
+  --maincolor: rgb(72, 61, 85);
+  --main2color: rgb(65, 56, 75);
+  --main3color: rgb(52, 44, 60);
+
+  /* --maincolor plus --maintint equals --main2color */
+  --maintint: rgba(0, 0, 0, 0.1);
+
+  /* --maincolor plus --main2tint equals --main3color */
+  /* main shadow color also */
+  --main2tint: rgba(0, 0, 0, 0.275);
+
+  /* Blue shades */
+  --blue-1: rgb(173, 178, 242);
+  --blue-2: rgb(129, 135, 216);
+  --blue-3: rgb(95, 102, 197);
+  --blue-4: rgb(63, 71, 167);
+  --blue-5: rgb(41, 48, 141);
+
+  /* Red shades */
+  --red-1: rgb(242, 173, 173);
+  --red-2: rgb(215, 128, 128);
+  --red-3: rgb(195, 93, 93);
+  --red-4: rgb(163, 59, 59);
+  --red-5: rgb(137, 37, 37);
+
+  font-size: 20px;
+  color: white;
+}
+
+p {
+  font-size: 1rem;
+  margin: 0%;
+}
+
+h1 {
+  font-size: 3rem;
+  margin: 0%;
+}
+
+h2 {
+  font-size: 2rem;
+  margin: 0%;
+}
+
+h3 {
+  font-size: 1.5rem;
+  margin: 0%;
+}
+
+hr {
+  height: 0px;
+}
+
+
+
+/* App.css */
+.app-container {
+  font-family: "Josefin Sans", sans-serif;
+  display: flex;
+  flex-direction: column;
+  background-color: var(--maincolor);
+}
+
+/* Header Styling */
+.top-header {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: 75px;
+  z-index: 5;
+}
+
+.top-header h1 {
+  text-align: center;
+  flex-grow: 1;
+  margin: 0;
+  color: white;
+  transform: translateY(4px);
+  z-index: 5;
+}
+
+.e-m-title {
+  padding-inline: 100px;
+  height: 100%;
+  display: flex;
+  align-items: center;
+}
+
+.left-box {
+  height: 100%;
+  width: 100%;
+  background-color: var(--blue-5);
+}
+
+.left-curve {
+  z-index: 10;
+  height: 75px;
+  min-width: 81px;
+  fill: var(--blue-5);
+  filter: 
+  drop-shadow(15px 0px 0px var(--blue-4))
+  drop-shadow(20px 0px 0px var(--blue-3))
+  drop-shadow(25px 0px 0px var(--blue-2))
+  drop-shadow(30px 0px 0px var(--blue-1));
+  transition: 0.5s;
+}
+
+.top-header:hover .left-curve {
+  filter:
+  drop-shadow(30px 0px 0px var(--blue-4))
+  drop-shadow(30px 0px 0px var(--blue-3))
+  drop-shadow(30px 0px 0px var(--blue-2))
+  drop-shadow(30px 0px 0px var(--blue-1));
+  transition: 0.5s;
+}
+
+.right-box {
+  height: 100%;
+  width: 100%;
+  background-color: var(--red-5);
+}
+
+.right-curve {
+  z-index: 10;
+  height: 75px;
+  min-width: 81px;
+  fill: var(--red-5);
+  filter: 
+  drop-shadow(-15px 0px 0px var(--red-4))
+  drop-shadow(-20px 0px 0px var(--red-3))
+  drop-shadow(-25px 0px 0px var(--red-2))
+  drop-shadow(-30px 0px 0px var(--red-1));
+  transition: 0.5s;
+}
+
+.top-header:hover .right-curve {
+  filter:
+  drop-shadow(-30px 0px 0px var(--red-4))
+  drop-shadow(-30px 0px 0px var(--red-3))
+  drop-shadow(-30px 0px 0px var(--red-2))
+  drop-shadow(-30px 0px 0px var(--red-1));
+  transition: 0.5s;
+}
+
+.nav-header {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 50px;
+  padding: 0%;
+  background: var(--main2color);
+  box-shadow: 0px 0px 15px 5px var(--main2tint) inset;
+  z-index: 6;
+}
+
+.nav-button {
+  height: 50px;
+  padding: 0 45px;
+  font-family: "Josefin Sans", sans-serif;
+  background-color: transparent;
+  border: none;
+  display: flex;
+  align-items: center;
+  transition: 0.15s;
+}
+
+.nav-button:hover {
+  box-shadow: inset 0px 0px 15px var(--main2tint);
+  transition: 0.15s;
+}
+
+.nav-button p {
+  font-size: 1.25rem;
+  transform: translateY(2px);
+  color: white;
+  font-weight: 500;
+  text-decoration: none;
+}
+
+a {
+  text-decoration: none;
+}
+
+button {
+  background: linear-gradient(to bottom, var(--maincolor), var(--main2color));
+  box-shadow: 0px 3px 1px var(--main2tint),
+  0px 1px 1px rgba(255, 255, 255, 0.1) inset;
+  padding-inline: 15px;
+  height: 2rem;
+  border: 0;
+  border-radius: 15px;
+  font-family: "Josefin Sans", sans-serif;
+  font-size: 1rem;
+  color: white;
+}
+
+button:hover {
+  background: linear-gradient(to bottom, var(--maincolor), var(--main3color));
+  box-shadow: 0px 3px 1px var(--main2tint) inset,
+  0px 1px 1px rgba(255, 255, 255, 0.1);
+  transform: translateY(3px);
+}
+
+button:active {
+  background: linear-gradient(to bottom, var(--main2color), var(--main3color));
+  box-shadow: 0px 3px 1px var(--main2tint) inset,
+  0px 1px 1px rgba(255, 255, 255, 0.1);
+}
+
+/* Main Content Area */
+.main-content {
+  flex: 1;
+  min-height: 100vh;
+}
+
+/* Decorative Elements */
+.stripe {
+  background-image:
+      repeating-linear-gradient(-45deg,
+          transparent,
+          transparent 1rem,
+          #00000015 1rem,
+          #00000010 2rem);
+  background-size: 150vw 150vh;
+  animation: barberpole 100s linear infinite;
+}
+
+@keyframes barberpole {
+  100% {
+      background-position: 100% 100%;
+      backdrop-filter: blur(5px);
+  }
+}
