+++ src/mapComponents/components.css
@@ -1,0 +1,151 @@
+.map-viewbox {
+  border: 3px white dotted;
+  width: calc(100% - 46px);
+  height: 50%;
+  padding: 20px;
+  border-radius: 15px;
+  aspect-ratio: 1.05;
+}
+
+.model-select {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: center;
+  text-align: center;
+}
+
+.model-select h2 {
+  margin-bottom: 0;
+}
+
+.model-select .button-container {
+  border-top: 3px white dotted;
+  width: 90%;
+  margin-bottom: 10px;
+  padding-top: 5px;
+}
+
+.model-select .button-container button {
+  font-size: 1rem;
+  padding: 5px 15px;
+  margin: 5px;
+}
+
+.model-select #pres {
+  display: flex;
+  flex-direction: column;
+  width: 88%;
+}
+
+.model-select #model button,
+.model-select #results button {
+  width: 45%;
+}
+
+.wincall-select {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: center;
+  text-align: center;
+  gap: 15px;
+}
+
+.wincall-select select {
+  background: linear-gradient(to bottom, var(--maincolor), var(--main2color));
+  width: 50%;
+  align-items: center;
+  justify-content: center;
+  display: flex;
+  height: 2rem;
+  border: 0;
+  border-radius: 15px;
+  box-shadow: 0px -3px 1px var(--main2tint) inset,
+    0px 1px 1px rgba(255, 255, 255, 0.1) inset;
+  font-family: "Josefin Sans", sans-serif;
+  font-size: 1.5rem;
+  font-weight: 600;
+  color: white;
+}
+
+.wincall-select h3 {
+  font-size: 1.25rem;
+}
+
+.call-button-box {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+}
+
+.call-button-box button {
+  width: 50%;
+  align-items: center;
+  justify-content: center;
+  display: flex;
+  height: 3rem;
+  border: 0;
+}
+
+.call-button-box .callD {
+  background: linear-gradient(to bottom, var(--blue-4), var(--blue-5));
+  margin-right: 5px;
+}
+
+.call-button-box .callR {
+  background: linear-gradient(to bottom, var(--red-4), var(--red-5));
+  margin-left: 5px;
+}
+
+.call-button-box .callD:hover {
+  background: linear-gradient(to bottom, var(--blue-5), var(--blue-5));
+}
+
+.call-button-box .callR:hover {
+  background: linear-gradient(to bottom, var(--red-5), var(--red-5));
+}
+
+.call-button-box button p {
+  font-family: "Josefin Sans", sans-serif;
+  font-size: 1.5rem;
+  font-weight: 600;
+  color: white;
+}
+
+.call-button-box button:hover p {
+  transform: translateY(3px);
+}
+
+.winput-number {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+  gap: 10px;
+  padding-block: 10px;
+}
+
+.winput-number input {
+  background: linear-gradient(to bottom, var(--maincolor), var(--main2color));
+  box-shadow: 0px 3px 1px var(--main2tint),
+  0px 1px 1px rgba(255, 255, 255, 0.1) inset;
+  width: 50%;
+  padding-inline: 15px;
+  height: 2rem;
+  border: 0;
+  border-radius: 15px;
+  font-family: "Josefin Sans", sans-serif;
+  font-size: 1.5rem;
+  font-weight: 600;
+  color: white;
+}
+
+.winput-number button {
+  width: 50%;
+}
+
