+++ src/pages/pages.css
@@ -1,0 +1,255 @@
+.menu-panel {
+  border-radius: 15px;
+  padding: 15px;
+  box-shadow: 0px 5px 2px var(--main2tint),
+  0px 1px 1px rgba(255, 255, 255, 0.1) inset;
+  width: calc(100% - 50px);
+  margin: 10px;
+  height: auto;
+  background-color: var(--maincolor);
+}
+
+.inset-panel {
+  border-radius: 15px;
+  padding: 15px;
+  box-shadow: 0px 5px 1px var(--main2tint) inset,
+    0px -1px 5px rgba(255, 255, 255, 0.1);
+  width: calc(100% - 50px);
+  margin: 10px;
+  height: auto;
+  background-color: var(--maintint);
+  backdrop-filter: blur(3px);
+}
+
+.inside-panel {
+  width: calc(100% - 30px);
+  margin: 0px;
+}
+
+.aboutText {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  justify-content: center;
+  align-self: baseline;
+}
+
+.legacyButton {
+  display: flex;
+  width: 100%;
+  justify-content: center;
+  align-self: center;
+  padding-block: 20px;
+}
+
+.swapPage {
+  background-color: var(--main2tint);
+  border-radius: 15px 15px 0px 0px;
+  display: flex;
+  gap: 15px;
+  padding: 5px 15px 10px 15px;
+  width: 320px;
+  height: 40px;
+  position: fixed;
+  z-index: 1000;
+  left: calc(50% - 160px);
+  bottom: 0px;
+  justify-content: center;
+  opacity: 50%;
+  transition: 0.5s;
+}
+
+.exitLegacy:hover,
+.swapPage:hover {
+  opacity: 100%;
+  transition: 0.15s;
+}
+
+.exitLegacy {
+  background-color: var(--main2tint);
+  border-radius: 15px 0px 0px 0px;
+  display: flex;
+  gap: 15px;
+  padding: 5px 5px 10px 15px;
+  width: 150px;
+  height: 40px;
+  position: fixed;
+  z-index: 1000;
+  right: 0px;
+  bottom: 0px;
+  justify-content: center;
+  opacity: 50%;
+  transition: 0.5s;
+}
+
+.legacyPortal {
+  transform: translateY(-130px);
+  width: 100%;
+  height: 100vh;
+  border: 0;
+  position: absolute;
+  z-index: 500;
+}
+
+/* Maps layout */
+.results {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-wrap: wrap;
+  gap: 15px;
+}
+
+.results-bar {
+  background-color: var(--main2tint);
+  border: 3px white outset;
+  border-radius: 15px;
+  height: 50px;
+  width: calc(100% - 56px);
+  display: flex;
+}
+
+.results-text {
+  width: 100%;
+  display: flex;
+  justify-content: space-around;
+  flex-direction: row;
+
+}
+
+.results-text h4 {
+  margin: 0;
+  margin-right: 1rem;
+}
+
+.results-text div {
+  display: flex;
+}
+
+.maps-columns {
+  display: flex;
+  height: 100%;
+}
+
+.maps-columns .left-column {
+  flex: 0 0 20%;
+}
+
+.maps-columns .right-column {
+  flex: 0 0 15%;
+  background-color: var(--maintint);
+}
+
+.maps-columns .left-column,
+.maps-columns .right-column {
+  flex-direction: column;
+  align-items: center;
+  display: flex;
+}
+
+.maps-columns .center-column {
+  flex: 1;
+  background-color: var(--maincolor);
+  box-shadow: 0px 0px 5px var(--main2tint);
+  display: flex;
+  padding: 10px;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: center;
+  gap: 15px;
+}
+
+
+/* Polls/Graphs Layout */
+.two-column-layout {
+  display: flex;
+  height: 100%;
+}
+
+.two-column-layout .left-column {
+  flex: 0 0 40%;
+  background-color: var(--main2color);
+}
+
+.two-column-layout .right-column {
+  flex: 1;
+  background-color: var(--maincolor);
+  padding: 20px;
+  box-shadow: 0px 0px 5px var(--main2tint);
+}
+
+/* About Layout */
+.about-columns {
+  display: flex;
+  height: 100%;
+}
+
+.about-columns .left-column {
+  flex: 0 0 30%;
+  background-color: var(--main2color);
+}
+
+.about-columns .center-column {
+  flex: 1;
+  background-color: var(--maincolor);
+  padding: 20px;
+  box-shadow: 0px 0px 5px var(--main2tint);
+  overflow-y: scroll;
+  height: calc(100vh - 170px);
+}
+
+.about-columns .right-column {
+  flex: 0 0 30%;
+  background-color: var(--main2color);
+}
+
+/* Responsive Adjustments at 960px */
+@media (max-width: 960px) {
+
+  /* Maps Page Responsive Layout*/
+  .maps-columns {
+    flex-direction: column;
+  }
+
+  .maps-columns .center-column {
+    order: 1;
+    padding: 20px;
+  }
+
+  .maps-columns .left-column {
+    order: 2;
+  }
+
+  .maps-columns .right-column {
+    order: 3;
+  }
+
+  /* Polls and Graphs Responsive Layout */
+  .two-column-layout {
+    flex-direction: column;
+  }
+
+  .two-column-layout .left-column {
+    height: calc(50% - 20px);
+    overflow-y: auto;
+  }
+
+  .two-column-layout .right-column {
+    height: 50%;
+  }
+
+  /* About Page Responsive Layout */
+  .about-columns {
+    display: flex;
+  }
+
+  .about-columns .left-column,
+  .about-columns .right-column {
+    display: none;
+  }
+
+  .about-columns .center-column {
+    width: 100%;
+  }
+}
